--- src/main/java/com/example/crudproject/CrudProjectApplication.java
@@ -1,13 +1,0 @@
-package com.example.crudproject;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class CrudProjectApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(CrudProjectApplication.class, args);
-	}
-
-}
