+++ src/main/java/com/example/crudproject/auth/CrudProjectApplication.java
@@ -1,0 +1,13 @@
+package com.example.crudproject.auth;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class CrudProjectApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(CrudProjectApplication.class, args);
+	}
+
+}
