+++ src/main/java/com/example/crudproject/auth/dto/JwtRequestDto.java
@@ -1,0 +1,4 @@
+package com.example.crudproject.auth.dto;
+
+public class JwtRequestDto {
+}
