+++ src/main/java/com/example/crudproject/auth/entity/Article.java
@@ -1,0 +1,40 @@
+package com.example.crudproject.auth.entity;
+
+import jakarta.persistence.*;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+import java.util.ArrayList;
+import java.util.List;
+
+@Getter
+@Entity
+@NoArgsConstructor
+public class Article {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @Setter
+    private String title;
+    @Setter
+    private String content;
+
+    @Setter
+    @ManyToOne
+    private Board board;
+    @Setter
+    @ManyToOne
+    private User user;
+
+    @OneToMany(mappedBy = "article")
+    private final List<Comment> commentList = new ArrayList<>();
+
+    public Article(String title, String content, Board board, User user) {
+        this.title = title;
+        this.content = content;
+        this.board = board;
+        this.user = user;
+    }
+}
