+++ src/main/java/com/example/crudproject/auth/filters/JwtTokenFilter.java
@@ -1,0 +1,4 @@
+package com.example.crudproject.auth.filters;
+
+public class JwtTokenFilter {
+}
