+++ src/main/java/com/example/crudproject/auth/jwt/JwtTokenUtils.java
@@ -1,0 +1,4 @@
+package com.example.crudproject.auth.jwt;
+
+public class JwtTokenUtils {
+}
