+++ src/main/java/com/example/crudproject/auth/repository/CommentRepository.java
@@ -1,0 +1,8 @@
+package com.example.crudproject.auth.repository;
+
+import com.example.crudproject.auth.entity.Comment;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface CommentRepository extends JpaRepository<Comment, Long> { }
