--- src/main/java/com/example/crudproject/repository/ArticleRepository.java
@@ -1,17 +1,0 @@
-
-
-package com.example.crudproject.repository;
-
-import com.example.crudproject.entity.Article;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.List;
-import java.util.Optional;
-
-@Repository
-public interface ArticleRepository extends JpaRepository<Article, Long> {
-    List<Article> findByTitleContaining(String title);
-    List<Article> findByContentContaining(String content);
-
-}
