--- src/main/java/com/example/crudproject/repository/BoardRepository.java
@@ -1,14 +1,0 @@
-package com.example.crudproject.repository;
-
-import com.example.crudproject.entity.Board;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.Optional;
-
-@Repository
-public interface BoardRepository extends JpaRepository<Board, Long> {
-
-    Optional<Board> findByCategory(String category);
-    boolean existsByCategory(String category);
-}
