--- src/main/java/com/example/crudproject/repository/CommentRepository.java
@@ -1,8 +1,0 @@
-package com.example.crudproject.repository;
-
-import com.example.crudproject.entity.Comment;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface CommentRepository extends JpaRepository<Comment, Long> { }
