+++ src/main/resources/templates/index.html
@@ -1,0 +1,23 @@
+<!DOCTYPE html>
+<html
+        lang="en"
+        xmlns:sec="http://www.w3.org/1999/xhtml"
+>
+<head>
+  <meta charset="UTF-8">
+  <title>Title</title>
+</head>
+<body>
+<h1>HOME</h1>
+<div sec:authorize="isAnonymous()">
+  <a href="/login">로그인</a>
+  <a href="/register">회원가입</a>
+</div>
+<div sec:authorize="isAuthenticated()">
+  <h3>반갑습니다, <span sec:authentication="name"></span>님!</h3>
+  <form action="/logout" method="post">
+    <input type="submit" value="로그아웃">
+  </form>
+</div>
+</body>
+</html>
