+++ src/main/resources/templates/login-form.html
@@ -1,0 +1,45 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+  <meta charset="UTF-8">
+  <title>Title</title>
+  <meta name="viewport" content="width=device-width, initial-scale=1.0">
+  <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.2.0/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-gH2yIJqKdNHPEq0n4Mqa/HGKIhSkIHeL5AyhkYV8i59U5AR6csBvApHHNl/vI1Bx" crossorigin="anonymous">
+</head>
+<body>
+<!--  <h1>Sign In</h1>-->
+<!--  <form th:action="@{/users/sign-in}" method="post">-->
+<!--    <input type="text" name="username" placeholder="아이디">-->
+<!--    <input type="password" name="password" placeholder="비밀번호">-->
+<!--    <button type="submit">로그인</button>-->
+<!--  </form>-->
+<main class="flex-shrink-0">
+  <section class="py-5">
+    <div class="container px-5">
+      <!-- login form-->
+      <div class="bg-light rounded-3 py-5 px-4 px-md-5 mb-5">
+        <div class="row gx-5 justify-content-center">
+          <div class="col-lg-8 col-xl-6">
+            <h1 class="text-center mb-5">로그인</h1>
+            <form action="/login" method="post">
+              <div class="form-floating mb-3">
+                <input class="form-control" id="identifier" name="username" type="text" placeholder="Enter your username...">
+                <label for="identifier">ID</label>
+              </div>
+              <div class="form-floating mb-3">
+                <input class="form-control" id="password" name="password" type="password" placeholder="Enter your password...">
+                <label for="password">Password</label>
+              </div>
+              <div class="d-grid"><button class="btn btn-primary btn-lg" id="sign-in-button" type="submit">Submit</button></div>
+              <div style="margin-top: 16px; text-align: right"><a href="/register">회원가입</a></div>
+            </form>
+          </div>
+        </div>
+
+      </div>
+    </div>
+  </section>
+</main>
+<script src="https://cdn.jsdelivr.net/npm/bootstrap@5.2.0/dist/js/bootstrap.bundle.min.js" integrity="sha384-A3rJD856KowSb7dwlZdYEkO39Gagi7vIsF0jrRAoQmDKKtQBHUuLZ9AsSv4jD4Xa" crossorigin="anonymous"></script>
+</body>
+</html>
