+++ src/main/resources/templates/my-profile.html
@@ -1,0 +1,14 @@
+<!DOCTYPE html>
+<html lang="en" xmlns:sec="http://www.w3.org/1999/xhtml">
+<head>
+  <meta charset="UTF-8">
+  <title>Title</title>
+</head>
+<body>
+  <h1>My Profile</h1>
+  <h3>반갑습니다. <span sec:authentication="name" ></span> </h3>
+  <form action="/logout">
+    <input type="submit" value="로그아웃">
+  </form>
+</body>
+</html>
