+++ src/main/resources/templates/register-form.html
@@ -1,0 +1,43 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+  <meta charset="UTF-8">
+  <title>Title</title>
+  <meta name="viewport" content="width=device-width, initial-scale=1.0">
+  <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.2.0/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-gH2yIJqKdNHPEq0n4Mqa/HGKIhSkIHeL5AyhkYV8i59U5AR6csBvApHHNl/vI1Bx" crossorigin="anonymous">
+</head>
+<body>
+<main class="flex-shrink-0">
+  <section class="py-5">
+    <div class="container px-5">
+      <!-- login form-->
+      <div class="bg-light rounded-3 py-5 px-4 px-md-5 mb-5">
+        <div class="row gx-5 justify-content-center">
+          <div class="col-lg-8 col-xl-6">
+            <h1 class="text-center mb-5">회원가입</h1>
+            <form action="/register" method="post">
+              <div class="form-floating mb-3">
+                <input class="form-control" id="identifier" name="username" type="text" placeholder="Enter your identifier...">
+                <label for="identifier">ID</label>
+              </div>
+              <div class="form-floating mb-3">
+                <input class="form-control" id="password" name="password" type="password" placeholder="Enter your password...">
+                <label for="password">Password</label>
+              </div>
+              <div class="form-floating mb-3">
+                <input class="form-control" id="password-check" name="password-check" type="password" placeholder="Re-enter your password...">
+                <label for="password-check">Password Check</label>
+              </div>
+              <div class="d-grid"><button class="btn btn-primary btn-lg" id="sign-in-button" type="submit">Submit</button></div>
+              <div style="margin-top: 16px; text-align: right"><a href="/users/login">로그인</a></div>
+            </form>
+          </div>
+        </div>
+
+      </div>
+    </div>
+  </section>
+</main>
+<script src="https://cdn.jsdelivr.net/npm/bootstrap@5.2.0/dist/js/bootstrap.bundle.min.js" integrity="sha384-A3rJD856KowSb7dwlZdYEkO39Gagi7vIsF0jrRAoQmDKKtQBHUuLZ9AsSv4jD4Xa" crossorigin="anonymous"></script>
+</body>
+</html>
